Makefile: use -q for texcount
I thought -nc would be sufficient, but alas, it's not: if one of
the source tex files causes texcount to report an error, the output
is not suitable for including IN tex, and building the thesis PDF
will fail with inscrutable errors like

    ! Package inputenc Error: Unicode character ^^[ (U+001B)
    (inputenc)                not set up for use with LaTeX.

    See the inputenc package documentation for explanation.
    Type  H <return>  for immediate help.
     ...

    l.24 \makewordcount
                       ^^M

The cause in this case is a wordcount.summary containing this
(rendered to not reproduce the funny glyph directly):

    !!! Environment ended while waiting for ^[nd{code}. !!!

where '^[' is U+001B ESCAPE

With -q, despite texcount's documentation dissuading its use, the output
still indicates that there were problems (with e.g. '(errors:15)') and
the command still prints out the specific errors on stderr, so it's not
entirely lost.