clean up standalone rules
The phony clean and purge rules were incorrectly specified in terms
of the source files, not the destinations (e.g. expanding to
"clean/chapter1-standalone.tex").

Avoid defining two variables STANDALONE and standalone; use
STANDALONE_SOURCES for the former and STANDALONE for the latter.

Define phone 'clean-standalone' and 'purge-standalone' rules; group
them with the phony 'standalone' rule. Depend upon these later
instead of  doing addprefix calculations in the dependency list for
clean/purge.

declare all three .PHONY within this section of the Makefile so all
the standalone stuff is together; remove the subsequent declaration
in the bottom-most .PHONY.

Signed-off-by: Jonathan Dowland <jon@dow.land>